chore: Move models endpoint to the endpoints section of app
