+++ webapp/app.py
@@ -27,6 +27,7 @@ from webapp.tutorials.views import init_tutorials
 from webapp.packages.store_packages import store_packages
 from webapp.endpoints.views import endpoints
 from webapp.endpoints.signing_keys import signing_keys
+from webapp.endpoints.models import models
 
 
 TALISKER_WSGI_LOGGER = logging.getLogger("talisker.wsgi")
@@ -71,6 +72,7 @@ def create_app(testing=False):
     app.register_blueprint(admin)
     app.register_blueprint(endpoints)
     app.register_blueprint(signing_keys)
+    app.register_blueprint(models)
     init_docs(app, "/docs")
     init_blog(app, "/blog")
     init_tutorials(app, "/tutorials")
